move public config structs to types crate
