+++ client/Cargo.toml
@@ -15,3 +15,4 @@ surf-disco = { workspace = true }
 tokio = { workspace = true }
 tracing = { workspace = true }
 vbs = { workspace = true }
+serde_json = { workspace = true }
