+++ types/Cargo.toml
@@ -56,6 +56,7 @@ tower-service = { version = "0.3", default-features = false }
 tracing = { workspace = true }
 url = { workspace = true }
 vbs = { workspace = true }
+vec1 = { workspace = true }
 
 [dev-dependencies]
 espresso-types = { path = ".", features = [ "testing" ] }
