+++ types/src/v0/v0_1/mod.rs
@@ -4,6 +4,7 @@ pub const VERSION: Version = Version { major: 0, minor: 1 };
 
 mod block;
 mod chain_config;
+mod config;
 mod fee_info;
 mod header;
 mod instance_state;
@@ -14,6 +15,7 @@ mod transaction;
 
 pub use block::*;
 pub use chain_config::*;
+pub use config::*;
 pub use fee_info::*;
 pub use header::Header;
 pub use instance_state::*;
