+++ types/src/v0/v0_3/mod.rs
@@ -7,7 +7,8 @@ pub use super::v0_1::{
     FeeMerkleTree, Header, Index, Iter, L1BlockInfo, L1Client, L1ClientOptions, L1Snapshot,
     NamespaceId, NsIndex, NsIter, NsPayload, NsPayloadBuilder, NsPayloadByteLen, NsPayloadOwned,
     NsPayloadRange, NsProof, NsTable, NsTableBuilder, NsTableValidationError, NumNss, NumTxs,
-    NumTxsRange, NumTxsUnchecked, Payload, PayloadByteLen, ResolvableChainConfig, TimeBasedUpgrade,
+    NumTxsRange, NumTxsUnchecked, Payload, PayloadByteLen, PublicHotShotConfig,
+    PublicNetworkConfig, PublicValidatorConfig, ResolvableChainConfig, TimeBasedUpgrade,
     Transaction, TxIndex, TxIter, TxPayload, TxPayloadRange, TxProof, TxTableEntries,
     TxTableEntriesRange, Upgrade, UpgradeMode, UpgradeType, ViewBasedUpgrade,
     BLOCK_MERKLE_TREE_HEIGHT, FEE_MERKLE_TREE_HEIGHT, NS_ID_BYTE_LEN, NS_OFFSET_BYTE_LEN,
